Move rayonix functions and constants into serialtbx
